requirements.txt
crispy-bootstrap5
Cython
dj-database-url
Django
django-cors-headers
django-crispy-forms
djangorestframework
pillow
psycopg
psycopg2-binary
python-dotenv
sqlparse
whitenoise
sslcommerz-lib